# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QuizSubmitRequest": {
        "additionalProperties": false,
        "properties": {
          "answers": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Answers",
            "type": "array"
          },
          "child_id": {
            "title": "Child Id",
            "type": "integer"
          }
        },
        "required": [
          "child_id",
          "answers"
        ],
        "title": "QuizSubmitRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DmkWords API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/miniapp/growth/summary": {
      "get": {
        "operationId": "growth_summary_api_miniapp_growth_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Growth Summary",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
    "/api/miniapp/leaderboard": {
      "get": {
        "description": "五榜单（周期榜仅有效会员可见；总榜含历史学员）。",
        "operationId": "leaderboard_api_miniapp_leaderboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "week",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "child_id",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Leaderboard",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
...
    "/api/miniapp/passport": {
      "get": {
        "operationId": "passport_api_miniapp_passport_get",
        "parameters": [
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Passport",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
    "/api/miniapp/points": {
      "get": {
        "operationId": "points_ledger_api_miniapp_points_get",
        "parameters": [
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Points Ledger",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
    "/api/miniapp/quiz/status-batch": {
      "get": {
        "description": "T43（U2）：书架角标批量状态（3 次 IN 查询禁 N+1；book_ids 逗号分隔≤50）。",
        "operationId": "quiz_status_batch_api_miniapp_quiz_status_batch_get",
        "parameters": [
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "book_ids",
            "required": true,
            "schema": {
              "title": "Book Ids",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quiz Status Batch",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
    "/api/miniapp/quiz/{book_id}": {
      "get": {
        "operationId": "get_quiz_api_miniapp_quiz__book_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "title": "Book Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quiz",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
    "/api/miniapp/quiz/{book_id}/submit": {
      "post": {
        "operationId": "submit_quiz_api_miniapp_quiz__book_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "title": "Book Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Quiz",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
...
    "/api/miniapp/reports/{kind}": {
      "get": {
        "description": "周报/月报数据（家长预览）。R-313：未缴费禁；过期/退会只读。",
        "operationId": "report_api_miniapp_reports__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
    "/api/miniapp/reports/{kind}/image": {
      "get": {
        "description": "周报/月报图片（query token：图片组件无法带头）。",
        "operationId": "report_image_api_miniapp_reports__kind__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "child_id",
            "required": true,
            "schema": {
              "title": "Child Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Image",
        "tags": [
          "growth-miniapp"
        ]
      }
    },
...
  }
}